Assignment1/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==24.1
requests==2.31.0
Pillow==10.1.0
